Move schedule logic into separate file
